+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
   domain ``s = 2i*pi*freq``.
 - New function ``utils.get_cell_numbers`` to get good values of number of cells
   for given primes.
+- Speed-up ``njitted.volume_average`` significantly thanks to @jcapriot.
 - Bugfix and workaround for the case where a ``sslsolver`` is used together
   with a provided initial ``efield``.